requirements.txt
wsproto==1.3.2
flask
flask-cors
flask-sqlalchemy
# psycopg2-binary # Opcional: Solo necesario si usas PostgreSQL
requests
# pywmapi # Nota: Asegúrate de que esta librería sea compatible con Python 3.11
cloudscraper
webdriver-manager
beautifulsoup4
gunicorn
python-dotenv
google-generativeai
